+++ howard/functions/commons.py
@@ -511,6 +511,7 @@ def find_nomen(
     transcripts: list = [],
     transcripts_source_order: list = None,
     pattern=None,
+    transcripts_len: int = None
 ) -> dict:
     """
     The function `find_nomen` takes a HGVS string and a list of transcripts, parses the HGVS string, and
@@ -570,25 +571,43 @@ def find_nomen(
     }
     nomen_dict = empty_nomen_dict.copy()
 
+    # Transcripts list
+    transcripts_list = transcripts.copy()
+
+    # Transcript as a list (create dict with rank, slow)
+    if isinstance(transcripts_list, list):
+        transcripts_list_tmp = {}
+        for rank, source in enumerate(transcripts_list, start=1):
+            transcripts_list_tmp[source] = rank
+        transcripts_list = transcripts_list_tmp
+
+
     # Pattern
     if pattern is None:
         pattern = "GNOMEN:TNOMEN:ENOMEN:CNOMEN:RNOMEN:NNOMEN:PNOMEN"
 
-    # Transcripts source order
+    # Transcripts source order and rank
     if transcripts_source_order is None:
         transcripts_source_order = ["column", "file"]
+    transcripts_source_order_rank = {}
+    for rank, source in enumerate(transcripts_source_order, start=1):
+        transcripts_source_order_rank[source] = rank
 
-    # Add transcripts list from file into source
-    transcripts_sources = {"file": transcripts}
+    # Transcripts file length
+    if transcripts_len is None:
+        transcripts_len = len(transcripts_list)
 
-    # Check if transcript in column transcript
+    # Check if transcript in column transcript, and add to transcripts list/rank
+    transcript_shift = 0
     if transcript is not None and str(transcript) != "nan":
-        transcripts_sources["column"] = str(transcript).split(",")
-
-    # Order transcript list by source order
-    transcripts_list = []
-    for order in transcripts_source_order:
-        transcripts_list += transcripts_sources.get(order, [])
+        transcript_list = str(transcript).split(",")
+        if transcripts_source_order_rank.get("column", 0) < transcripts_source_order_rank.get("file", 0):
+            transcript_shift = len(transcript_list)
+        for rank, transcript in enumerate(transcript_list, start=1):
+            if transcripts_source_order_rank.get("column", 0) < transcripts_source_order_rank.get("file", 0):
+                transcripts_list[transcript] = rank - len(transcript_list)
+            elif transcript not in transcripts:
+                transcripts_list[transcript] = rank + transcripts_len
 
     # If hgvs not empty
     if str(hgvs) != "nan":
@@ -629,11 +648,11 @@ def find_nomen(
                         or one_nomen_dict["TNOMEN"] in transcripts_list
                     ):
                         rank = max(
-                            get_index(one_nomen_dict["TVNOMEN"], transcripts_list),
-                            get_index(one_nomen_dict["TNOMEN"], transcripts_list),
-                        )
+                            transcripts_list.get(one_nomen_dict["TVNOMEN"], - transcript_shift - 1),
+                            transcripts_list.get(one_nomen_dict["TNOMEN"], - transcript_shift - 1)
+                            ) + transcript_shift
                         if rank >= 0:
-                            one_nomen_score += 100 * (len(transcripts_list) - rank)
+                            one_nomen_score += 100 * (len(transcripts_list) - rank + 1)
 
                 elif re.match(r"^[NX]P_(.*)$", one_hgvs_infos):
                     # Transcript Protein with version
@@ -2704,11 +2723,11 @@ def get_argument_to_mk(arg: str, argument: dict = {}, mode: str = "mk") -> str:
     else:
         text += "\n<small>\n"
         text += "\n> ```\n"
-        text += f"> {text_header}"
+        text += f">     {text_header}"
         text += "\n> \n> "
         text += str(help).strip().replace("\n", "\n> ")
         text += "\n> ```\n"
-        text += "</small>\n"
+        text += "\n</small>\n"
         text += "\n"
         text += ""